moved scopeOptions() to utils
