+++ content/docs/tableau-test/tableau-test.mdx
@@ -11,6 +11,9 @@ description: "Testing embedding techniques for the Tableau Component"
     },
     {
       url: "https://public.tableau.com/views/Greenlight/GreenlightProfitDashboardfixed",
+      layout: {
+        phone: {width: 375, height: 1000},
+      },
     },
     {url: "https://public.tableau.com/views/Greenlight/GreenlightProfitDashboardautomatic"},
     {url: "https://public.tableau.com/views/Greenlight/GreenlightProfitDashboardrange"},
