+++ src/components/viz/viz.module.css
@@ -8,7 +8,3 @@
 .vizDiv iframe {
   margin: auto;
 }
-
-.placeholder {
-  min-height: 750px;
-}
